Protect against DNS rebind attacks
We resolve the incoming hostname. If it passes, we let HTTP::Tiny run,
which resolves it again. An attacker that controls the DNS could use a
TTL of 0 and return a whitelisted IP first, then a blacklisted IP on the
second query, defeating the protection.

The solution is to only do one DNS query, and pass the results down to
HTTP::Tiny.